requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Free LLM Provider - Groq (Free Tier: https://console.groq.com)
# groq>=0.13.0 fixes the httpx 'proxies' TypeError on Windows
groq==0.13.0
httpx==0.27.2

# Free Search - DuckDuckGo (No API key needed)
duckduckgo-search==6.3.7
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.3

# Logging
loguru==0.7.2
